function_code/requirements.txt
pyyaml==6.0
google==3.0.0
google_cloud_storage==2.7.0